% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{Case}
\alias{Case}
\title{Define a case for \code{\link[=Match]{Match()}}}
\usage{
Case(condition, content)
}
\arguments{
\item{condition}{A reactive expression that returns a logical value.}

\item{content}{Tag tree to render when this case matches.}
}
\value{
A case definition (a list).
}
\description{
Define a case for \code{\link[=Match]{Match()}}
}
