% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{Default}
\alias{Default}
\title{Define a default (fallback) case for \code{\link[=Match]{Match()}}}
\usage{
Default(content)
}
\arguments{
\item{content}{Tag tree to render when no other case matches.}
}
\value{
A case definition (a list).
}
\description{
A convenience wrapper around \code{\link[=Case]{Case()}} with a condition that is always
\code{TRUE}. Place this as the last argument to \code{Match}.
}
