% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tags.R
\name{nacre_dependency}
\alias{nacre_dependency}
\title{nacre JavaScript dependency}
\usage{
nacre_dependency()
}
\value{
An \code{html_dependency} object.
}
\description{
Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} for the client-side nacre
runtime (\code{nacre.js}).
}
\keyword{internal}
