% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tags.R
\name{nacre_id_counter}
\alias{nacre_id_counter}
\title{Create a local ID counter for use within a single \code{process_tags} call}
\usage{
nacre_id_counter()
}
\value{
A function that returns the next ID each time it is called.
}
\description{
Create a local ID counter for use within a single \code{process_tags} call
}
\keyword{internal}
