% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tags.R
\name{nacre_next_id}
\alias{nacre_next_id}
\title{Generate a unique nacre element ID}
\usage{
nacre_next_id()
}
\value{
A character string.
}
\description{
Returns an auto-incrementing ID of the form \code{"nacre-1"}, \code{"nacre-2"}, etc.
}
\keyword{internal}
