% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{nacreOutput}
\alias{nacreOutput}
\title{Create a nacre UI output placeholder}
\usage{
nacreOutput(id)
}
\arguments{
\item{id}{The output ID, matching the corresponding \code{renderNacre} call.}
}
\value{
An HTML tag with the nacre dependency.
}
\description{
Creates a \code{\link[shiny:htmlOutput]{shiny::uiOutput()}} with the nacre JavaScript dependency
attached. Use this in a standard Shiny UI to mark where \code{\link[=renderNacre]{renderNacre()}}
should inject its content.
}
