% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{PlotOutput}
\alias{PlotOutput}
\title{Embed a plot output in a nacre tag tree}
\usage{
PlotOutput(expr, ..., env = parent.frame())
}
\arguments{
\item{expr}{An expression that produces a plot.}

\item{...}{Additional arguments passed to \code{\link[shiny:plotOutput]{shiny::plotOutput()}}.}

\item{env}{The environment in which to evaluate \code{expr}.}
}
\value{
A nacre output node.
}
\description{
Shorthand for \code{Output(renderPlot, plotOutput, ...)}.
}
