% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{TableOutput}
\alias{TableOutput}
\title{Embed a table output in a nacre tag tree}
\usage{
TableOutput(expr, ..., env = parent.frame())
}
\arguments{
\item{expr}{An expression that produces a table.}

\item{...}{Additional arguments passed to \code{\link[shiny:renderTable]{shiny::tableOutput()}}.}

\item{env}{The environment in which to evaluate \code{expr}.}
}
\value{
A nacre output node.
}
\description{
Shorthand for \code{Output(renderTable, tableOutput, ...)}.
}
